Enhanced handling of apostrophes
If a label is used with MessageFormat, then single quotes in the text
must be represented by doubled single quotes '' throughout a String.
However, for translators it is not obvious in what context the label
is used.

When downloading from POEditor, this change doubles the single quotes
if the string contains a curly brace (which indicates that it is used
in MessageFormat).

When uploading to POEditor, it is removing double quotes.